Fix error extracting the number of discounts
Apparently always asking for thousands of discounts does not always
work, but when there are only a few discounts, the number is not
mentioned explicitly anywhere in the page.

The solution is to combine two approaches. If the number of discounts is
not found, try counting the number of discounts directly in the front
page.